move 'case_restore' code to OTA app
