+++ corehq/apps/ota/urls.py
@@ -10,6 +10,7 @@ from corehq.apps.ota.views import (
     restore,
     search,
     registry_case,
+    case_restore,
 )
 
 urlpatterns = [
@@ -24,4 +25,5 @@ urlpatterns = [
     url(r'^get_next_id/$', get_next_id, name='get_next_id'),
     url(r'^recovery_measures/(?P<build_id>[\w-]+)/$', recovery_measures, name='recovery_measures'),
     url(r'^registry_case/(?P<app_id>[\w-]+)/$', registry_case, name='registry_case'),
+    url(r'^case_restore/(?P<case_id>[\w\-]+)/$', case_restore, name='case_restore'),
 ]
